+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Also include container status in pod debugging output
+
 ## [1.27.0] - 2024-10-03
 
 ### Added